Much improved stack tracking - now implements all stack manipulation operations
Introduced JakStack which is an abstract class for dealing with information from the byte code stack.
JakTypeStack is the implementation used by JAK's optional type stack tracking.

+++ net.dougqh.jak/src/net/dougqh/jak/TypeWriter.java
@@ -279,7 +279,11 @@ final class TypeWriter {
 			srcDir,
 			this.className.replace( '.', '/' ) + ".class" );
 		
-		classFile.getParentFile().mkdirs();
+		File packageDir = classFile.getParentFile();
+		boolean success = packageDir.mkdirs();
+		if ( ! success && ! packageDir.exists() ) {
+			throw new IOException( "Unable to create directory " + packageDir );
+		}
 		
 		FileOutputStream out = new FileOutputStream( classFile );
 		try {
